+++ src/bot/Bot.ts
@@ -1,7 +1,7 @@
 import { Client, ClientEvents } from "discord.js";
 import { BotOptions, BotOptionsArgument, DefaultBotOptions } from "./Options";
 import { ClientEventNames, deepMerge, GuildMessage, isGuildMessage } from "../utils";
-import { Logger } from "../Logger";
+import { Logger } from "../logger/Logger";
 import { helpCommand } from "../builtIn";
 import { AnyCommand, Command, CommandContext, CommandStorage } from "../command";
 import { BotEventListener, createEventStorage, EmitOf, createNodeEmitterLink } from "../event";
