+++ src/logger/Log.ts
@@ -1,0 +1,9 @@
+/**
+ * Тип лога
+ */
+export type LogType =
+    | "log"
+    | "warning"
+    | "error"
+    | "task"
+    | "success";
