+++ src/logger/Logger.ts
@@ -1,49 +1,6 @@
-import { createEventStorage, EmitOf } from "./event";
-
-/**
- * Тип лога
- */
-export type LogType =
-    | "log"
-    | "warning"
-    | "error"
-    | "task"
-    | "success";
-
-/**
- * Тема логгера в консоли
- */
-export interface LoggerConsoleTheme {
-    /**
-     * @param logType тип лога
-     * @param log лог
-     * @param taskLevel уровень 'вложенности' лога (увеличивается при `task` и уменьшается при `endTask`)
-     * @param taskCompleted вызывается ли тема внутри `endTask`
-     * @returns новую строку лог, которая позже попадает в `console.log`
-     */
-    (logType: LogType, log: string, taskLevel: number, taskCompleted: boolean): string;
-}
-
-/**
- * Настройки логгера
- */
-export interface LoggerOptions {
-    /**
-     * Спрятать все логи в консоли (событие log всё ещё работает)
-     */
-    readonly hideInConsole: boolean;
-
-    /**
-     * Игнорировать ли предупреждения (`logger.warning`)
-     * @default false
-     */
-    readonly ignoreWarnings: boolean;
-
-    /**
-     * Функция 'темы' логгера. Вызывается перед `console.log`
-     */
-    readonly consoleTheme?: LoggerConsoleTheme;
-}
+import { createEventStorage, EmitOf } from "../event";
+import { LoggerConsoleTheme, LoggerOptions } from "./Options";
+import { LogType } from "./Log";
 
 export interface Logger extends LoggerOptions { }
